first pass of conversion of the box component
